Add cursor tests

+++ idb/src/cursor/key_cursor.rs
@@ -8,6 +8,7 @@ use crate::{utils::wait_request, CursorDirection, Error};
 #[derive(Debug, Clone, PartialEq, Eq)]
 pub struct KeyCursor {
     inner: SysKeyCursor,
+    finished: bool,
 }
 
 impl KeyCursor {
@@ -24,40 +25,106 @@ impl KeyCursor {
 
     /// Returns the key of the cursor. Returns an [`Error`] if the cursor is advancing or is finished.
     pub fn key(&self) -> Result<JsValue, Error> {
-        self.inner.key().map_err(Into::into)
+        if self.finished {
+            Ok(JsValue::null())
+        } else {
+            self.inner.key().map_err(Into::into)
+        }
     }
 
     /// Returns the effective key of the cursor. Returns an [`Error`] if the cursor is advancing or is finished.
     pub fn primary_key(&self) -> Result<JsValue, Error> {
-        self.inner.primary_key().map_err(Into::into)
+        if self.finished {
+            Ok(JsValue::null())
+        } else {
+            self.inner.primary_key().map_err(Into::into)
+        }
     }
 
     /// Advances the cursor through the next count records in range.
-    pub fn advance(&self, count: u32) -> Result<(), Error> {
-        self.inner.advance(count).map_err(Into::into)
+    pub async fn advance(&mut self, count: u32) -> Result<(), Error> {
+        if self.finished {
+            return Err(Error::CursorFinished);
+        }
+
+        let request = self.inner.request();
+
+        self.inner.advance(count)?;
+        let cursor: JsValue = wait_request(request).await?;
+
+        if cursor.is_null() {
+            self.finished = true;
+        } else {
+            let inner = SysKeyCursor::try_from(cursor)?;
+            self.inner = inner;
+        }
+
+        Ok(())
     }
 
     /// Advances the cursor to the next record in range matching or after key (if provided).
-    pub fn next(&self, key: Option<&JsValue>) -> Result<(), Error> {
-        self.inner.next(key).map_err(Into::into)
+    pub async fn next(&mut self, key: Option<&JsValue>) -> Result<(), Error> {
+        if self.finished {
+            return Err(Error::CursorFinished);
+        }
+
+        let request = self.inner.request();
+
+        self.inner.next(key)?;
+        let cursor: JsValue = wait_request(request).await?;
+
+        if cursor.is_null() {
+            self.finished = true;
+        } else {
+            let inner = SysKeyCursor::try_from(cursor)?;
+            self.inner = inner;
+        }
+
+        Ok(())
     }
 
     /// Advances the cursor to the next record in range matching or after key and primary key. Returns an [`Error`] if
     /// the source is not an [`Index`](crate::Index).
-    pub fn next_primary_key(&self, key: &JsValue, primary_key: &JsValue) -> Result<(), Error> {
-        self.inner
-            .next_primary_key(key, primary_key)
-            .map_err(Into::into)
+    pub async fn next_primary_key(
+        &mut self,
+        key: &JsValue,
+        primary_key: &JsValue,
+    ) -> Result<(), Error> {
+        if self.finished {
+            return Err(Error::CursorFinished);
+        }
+
+        let request = self.inner.request();
+
+        self.inner.next_primary_key(key, primary_key)?;
+        let cursor: JsValue = wait_request(request).await?;
+
+        if cursor.is_null() {
+            self.finished = true;
+        } else {
+            let inner = SysKeyCursor::try_from(cursor)?;
+            self.inner = inner;
+        }
+
+        Ok(())
     }
 
     /// Updated the record pointed at by the cursor with a new value.
     pub async fn update(&self, value: &JsValue) -> Result<JsValue, Error> {
+        if self.finished {
+            return Err(Error::CursorFinished);
+        }
+
         let request = self.inner.update(value)?;
         wait_request(request).await
     }
 
     /// Delete the record pointed at by the cursor with a new value.
     pub async fn delete(&self) -> Result<(), Error> {
+        if self.finished {
+            return Err(Error::CursorFinished);
+        }
+
         let request = self.inner.delete()?;
         let _: JsValue = wait_request(request).await?;
         Ok(())
@@ -66,7 +133,10 @@ impl KeyCursor {
 
 impl From<SysKeyCursor> for KeyCursor {
     fn from(inner: SysKeyCursor) -> Self {
-        Self { inner }
+        Self {
+            inner,
+            finished: false,
+        }
     }
 }
 
@@ -81,7 +151,10 @@ impl TryFrom<JsValue> for KeyCursor {
 
     fn try_from(value: JsValue) -> Result<Self, Self::Error> {
         let inner = value.try_into()?;
-        Ok(Self { inner })
+        Ok(Self {
+            inner,
+            finished: false,
+        })
     }
 }
